django-rest-docker-template/templates/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==38.0.1
distro==1.7.0
Django==4.1.1
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flake8==5.0.4
idna==3.4
jsonschema==3.2.0
mccabe==0.7.0
packaging==21.3
paramiko==2.11.0
psycopg2==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
texttable==1.6.4
tzdata==2022.4
urllib3==1.26.0
websocket-client==0.59.0